# asm/amd64: avoids map allocation per node